Updated Level Camera system

+++ src/main/java/entities/constants/EntityConstants.java
@@ -4,5 +4,5 @@ package entities.constants;
 //into an Entities object.
 public abstract class EntityConstants implements AnimationConstants, HitboxConstants{
 
-    public abstract int getSpeed();
+    public abstract float getSpeed();
 }
